+++ tests/testthat/test-test-app.R
@@ -1,0 +1,8 @@
+context("app-file")
+library(shinytest)
+test_that("analyzeBehaviorspace works", {
+  skip_on_cran()
+
+  appdir <- system.file(package="analyzeBehaviorspace", "analyzeBehaviorspace")
+  expect_pass(testApp(appdir, compareImages = FALSE))
+})